final_project/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-dateutil==2.8.2
pytz==2023.3
avro-python3==1.11.2
python-json-logger==2.0.7
schedule==1.2.0

